pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "road_runner"
version = "0.1.0"
description = "System-Level Test Execution Engine CLI"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "road_runner maintainers"}]
license = {text = "MIT"}
dependencies = [
    "typer[all]>=0.9.0",
    "PyYAML>=6.0",
    "jinja2>=3.1",
    "rich>=13.7",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
road_runner = "road_runner.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true
